Refactor FakeDataset into conftest.py
